ARCH: Move AI provider settings controller out of AppState
